requirements.txt
toml
rich~=13.3.5
pandas~=2.2.1
six
deepdiff
twine
setuptools~=69.2.0
tabulate~=0.9.0
requests~=2.31.0
python-dotenv~=1.0.1